requirements.txt
beautifulsoup4==4.8.0
Flask==1.1.1
Flask-Limiter==1.0.1
gunicorn==19.9.0
icalendar==4.0.3
lxml==4.4.1
pycryptodome==3.9.0
pytz==2019.2
requests-toolbelt==0.8.0
requests==2.18.4
